Jsonparser now remembers whether to remove parallel edges, and added a test.
